# ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "",
    "version": ""
  },
  "paths": {
...
    "/operations/{id}": {
      "delete": {
        "operationId": "cancel_operation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Operation id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
...
    "/pipelines": {
      "post": {
        "operationId": "start_pipeline",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartPipelineRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartPipelineResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "NodeId": {
        "type": "string",
        "format": "uuid"
      },
...
      "PageId": {
        "type": "string",
        "format": "uuid"
      },
...
      "ReadingOrder": {
        "type": "string",
        "enum": ["rtl", "ltr", "custom"]
      },
      "Region": {
        "type": "object",
        "required": ["x", "y", "width", "height"],
        "properties": {
          "height": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "width": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "x": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "y": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "StartPipelineRequest": {
        "type": "object",
        "required": ["steps"],
        "properties": {
          "boxPadding": {
            "type": ["number", "null"],
            "format": "float",
            "description": "Pixels to inset text from each edge of its layout box, to stop glyphs\nand strokes being clipped at the box border."
          },
          "defaultFont": {
            "type": ["string", "null"]
          },
          "defaultFontSize": {
            "type": ["number", "null"],
            "format": "float",
            "description": "Global render defaults applied when a text node has no explicit\nper-node override. The renderer falls back to auto-fit / predicted\nvalues when these are absent."
          },
          "pages": {
            "type": ["array", "null"],
            "items": {
              "$ref": "#/components/schemas/PageId"
            },
            "description": "`None` → whole project, `Some(pages)` → just those pages."
          },
          "readingOrder": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ReadingOrder"
              }
            ]
          },
          "region": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Region",
                "description": "Optional bounding-box hint for inpainter engines (repair-brush)."
              }
            ]
          },
          "shaderEffect": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TextShaderEffect"
              }
            ]
          },
          "shaderStroke": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TextStrokeStyle"
              }
            ]
          },
          "sourceLanguage": {
            "type": ["string", "null"],
            "description": "Language of the source text, as an OCR hint (e.g. \"Korean\").\n`None` = OCR model auto-detect."
          },
          "steps": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Engine ids (`inventory::submit!` ids) to run in order."
          },
          "systemPrompt": {
            "type": ["string", "null"]
          },
          "targetLanguage": {
            "type": ["string", "null"]
          },
          "textAlign": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TextAlign"
              }
            ]
          },
          "textNodeIds": {
            "type": ["array", "null"],
            "items": {
              "$ref": "#/components/schemas/NodeId"
            },
            "description": "Optional text-node ids for engines that can operate on individual blocks."
          }
        }
      },
      "StartPipelineResponse": {
        "type": "object",
        "required": ["operationId"],
        "properties": {
          "operationId": {
            "type": "string"
          }
        }
      },
      "TextAlign": {
        "type": "string",
        "enum": ["left", "center", "right"]
      },
...
      "TextShaderEffect": {
        "type": "object",
        "properties": {
          "bold": {
            "type": "boolean"
          },
          "italic": {
            "type": "boolean"
          }
        }
      },
      "TextStrokeStyle": {
        "type": "object",
        "properties": {
          "color": {
            "type": ["array", "null"],
            "items": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            },
            "description": "Explicit outline colour. `None` = automatic: contrast against the\nresolved text colour (white text gets a black outline and vice\nversa). Became `Option` in scene format v5; earlier formats stored a\nrequired colour that defaulted to white, converted on upgrade in\n`session.rs::compat`."
          },
          "enabled": {
            "type": "boolean"
          },
          "widthPx": {
            "type": ["number", "null"],
            "format": "float"
          }
        }
      },
...
    }
  }
}